+++ dot-lms/src/Components/AdminComponents/ShowCourse.js
@@ -160,13 +160,19 @@ const [open_edit, setOpen_edit] = React.useState(false);
       }
 
       function delete_chapter(chapter_id){
-        axios.delete("http://127.0.0.1:8000/api/Chapter"+chapter_id)
+        // eslint-disable-next-line no-restricted-globals
+        if (confirm("Are you Sure")) {
+          axios.delete("http://127.0.0.1:8000/api/Chapter"+chapter_id)
         .then(()=>{
          toast.success('Chapter deleted Successfully',{
            position:toast.POSITION.BOTTOM_CENTER
          })
          window.location.reload(true)
         })
+        } else {
+          console.log('cancled deletion')
+        }
+        
       }
